refactor: separate classes
